Fallback to RegexLikeMatcher for dynamic like pattern
The compilation time for DFA is costly compared
to regex. Therefore, for dynamic like pattern it
is better to use a regex-based matcher.

Benchmarks:
Before:
query: select count(shipmode) from lineitem where shipinstruct like concat('%', concat("shipmode", '%'));
time: 27:04 min

After:
query: select count(shipmode) from lineitem where shipinstruct like concat('%', concat("shipmode", '%'));
time: 4:44 sec
