+++ core/trino-main/src/test/java/io/trino/operator/scalar/BenchmarkLike.java
@@ -13,15 +13,11 @@
  */
 package io.trino.operator.scalar;
 
-import io.airlift.jcodings.specific.NonStrictUTF8Encoding;
-import io.airlift.joni.Matcher;
-import io.airlift.joni.Option;
-import io.airlift.joni.Regex;
-import io.airlift.joni.Syntax;
 import io.airlift.slice.Slice;
 import io.airlift.slice.Slices;
+import io.trino.likematcher.DfaLikeMatcher;
 import io.trino.likematcher.LikeMatcher;
-import io.trino.type.JoniRegexp;
+import io.trino.likematcher.RegexLikeMatcher;
 import org.openjdk.jmh.annotations.Benchmark;
 import org.openjdk.jmh.annotations.BenchmarkMode;
 import org.openjdk.jmh.annotations.Fork;
@@ -40,13 +36,6 @@ import org.openjdk.jmh.runner.options.VerboseMode;
 
 import java.util.Optional;
 
-import static io.airlift.joni.constants.MetaChar.INEFFECTIVE_META_CHAR;
-import static io.airlift.joni.constants.SyntaxProperties.OP_ASTERISK_ZERO_INF;
-import static io.airlift.joni.constants.SyntaxProperties.OP_DOT_ANYCHAR;
-import static io.airlift.joni.constants.SyntaxProperties.OP_LINE_ANCHOR;
-import static io.trino.spi.StandardErrorCode.INVALID_FUNCTION_ARGUMENT;
-import static io.trino.util.Failures.checkCondition;
-import static java.nio.charset.StandardCharsets.UTF_8;
 import static java.util.concurrent.TimeUnit.MILLISECONDS;
 import static java.util.concurrent.TimeUnit.NANOSECONDS;
 import static org.openjdk.jmh.annotations.Mode.AverageTime;
@@ -55,24 +44,11 @@ import static org.openjdk.jmh.annotations.Scope.Thread;
 @State(Thread)
 @OutputTimeUnit(NANOSECONDS)
 @BenchmarkMode(AverageTime)
-@Fork(3)
-@Warmup(iterations = 10, time = 500, timeUnit = MILLISECONDS)
-@Measurement(iterations = 30, time = 500, timeUnit = MILLISECONDS)
+@Fork(1)
+@Warmup(iterations = 2, time = 500, timeUnit = MILLISECONDS)
+@Measurement(iterations = 5, time = 500, timeUnit = MILLISECONDS)
 public class BenchmarkLike
 {
-    private static final Syntax SYNTAX = new Syntax(
-            OP_DOT_ANYCHAR | OP_ASTERISK_ZERO_INF | OP_LINE_ANCHOR,
-            0,
-            0,
-            Option.NONE,
-            new Syntax.MetaCharTable(
-                    '\\',                           /* esc */
-                    INEFFECTIVE_META_CHAR,          /* anychar '.' */
-                    INEFFECTIVE_META_CHAR,          /* anytime '*' */
-                    INEFFECTIVE_META_CHAR,          /* zero or one time '?' */
-                    INEFFECTIVE_META_CHAR,          /* one or more time '+' */
-                    INEFFECTIVE_META_CHAR));        /* anychar anytime */
-
     @State(Thread)
     public static class Data
     {
@@ -90,8 +66,8 @@ public class BenchmarkLike
 
         private Slice data;
         private byte[] bytes;
-        private JoniRegexp joniPattern;
-        private LikeMatcher matcher;
+        private RegexLikeMatcher joniRegexMatcher;
+        private DfaLikeMatcher dfaMatcher;
 
         @Setup
         public void setup()
@@ -108,87 +84,35 @@ public class BenchmarkLike
                         default -> throw new IllegalArgumentException("Unknown pattern: " + pattern);
                     });
 
-            matcher = LikeMatcher.compile(pattern, Optional.empty());
-            joniPattern = compileJoni(Slices.utf8Slice(pattern).toStringUtf8(), '0', false);
+            dfaMatcher = DfaLikeMatcher.compile(pattern, Optional.empty());
+            joniRegexMatcher = RegexLikeMatcher.compile(pattern, Optional.empty());
 
             bytes = data.getBytes();
         }
     }
 
     @Benchmark
-    public boolean benchmarkJoni(Data data)
+    public LikeMatcher benchmarkJoniRegexCompile(Data data)
     {
-        return likeVarchar(data.data, data.joniPattern);
+        return RegexLikeMatcher.compile(data.pattern, Optional.empty());
     }
 
     @Benchmark
-    public boolean benchmarkCurrent(Data data)
+    public LikeMatcher benchmarkDfaCompile(Data data)
     {
-        return data.matcher.match(data.bytes, 0, data.bytes.length);
+        return DfaLikeMatcher.compile(data.pattern, Optional.empty());
     }
 
-    public static boolean likeVarchar(Slice value, JoniRegexp pattern)
+    @Benchmark
+    public boolean benchmarkJoniRegexMatch(Data data)
     {
-        Matcher matcher;
-        int offset;
-        if (value.hasByteArray()) {
-            offset = value.byteArrayOffset();
-            matcher = pattern.regex().matcher(value.byteArray(), offset, offset + value.length());
-        }
-        else {
-            offset = 0;
-            matcher = pattern.matcher(value.getBytes());
-        }
-        return matcher.match(offset, offset + value.length(), Option.NONE) != -1;
+        return data.joniRegexMatcher.match(data.bytes, 0, data.bytes.length);
     }
 
-    private static JoniRegexp compileJoni(String patternString, char escapeChar, boolean shouldEscape)
+    @Benchmark
+    public boolean benchmarkDFAMatch(Data data)
     {
-        byte[] bytes = likeToRegex(patternString, escapeChar, shouldEscape).getBytes(UTF_8);
-        Regex joniRegex = new Regex(bytes, 0, bytes.length, Option.MULTILINE, NonStrictUTF8Encoding.INSTANCE, SYNTAX);
-        return new JoniRegexp(Slices.wrappedBuffer(bytes), joniRegex);
-    }
-
-    private static String likeToRegex(String patternString, char escapeChar, boolean shouldEscape)
-    {
-        StringBuilder regex = new StringBuilder(patternString.length() * 2);
-
-        regex.append('^');
-        boolean escaped = false;
-        for (char currentChar : patternString.toCharArray()) {
-            checkEscape(!escaped || currentChar == '%' || currentChar == '_' || currentChar == escapeChar);
-            if (shouldEscape && !escaped && (currentChar == escapeChar)) {
-                escaped = true;
-            }
-            else {
-                switch (currentChar) {
-                    case '%' -> {
-                        regex.append(escaped ? "%" : ".*");
-                        escaped = false;
-                    }
-                    case '_' -> {
-                        regex.append(escaped ? "_" : ".");
-                        escaped = false;
-                    }
-                    default -> {
-                        // escape special regex characters
-                        switch (currentChar) {
-                            case '\\', '^', '$', '.', '*' -> regex.append('\\');
-                        }
-                        regex.append(currentChar);
-                        escaped = false;
-                    }
-                }
-            }
-        }
-        checkEscape(!escaped);
-        regex.append('$');
-        return regex.toString();
-    }
-
-    private static void checkEscape(boolean condition)
-    {
-        checkCondition(condition, INVALID_FUNCTION_ARGUMENT, "Escape character must be followed by '%%', '_' or the escape character itself");
+        return data.dfaMatcher.match(data.bytes, 0, data.bytes.length);
     }
 
     public static void main(String[] args)
