+++ core/trino-main/src/test/java/io/trino/type/TestLikePatternType.java
@@ -13,7 +13,9 @@
  */
 package io.trino.type;
 
+import io.trino.likematcher.DfaLikeMatcher;
 import io.trino.likematcher.LikeMatcher;
+import io.trino.likematcher.RegexLikeMatcher;
 import io.trino.spi.block.Block;
 import io.trino.spi.block.BlockBuilder;
 import io.trino.spi.block.PageBuilderStatus;
@@ -27,11 +29,28 @@ import static org.assertj.core.api.Assertions.assertThat;
 public class TestLikePatternType
 {
     @Test
-    public void testGetObject()
+    public void testGetObjectForDfaLikeMatcher()
     {
         BlockBuilder blockBuilder = LIKE_PATTERN.createBlockBuilder(new PageBuilderStatus().createBlockBuilderStatus(), 10);
-        LIKE_PATTERN.writeObject(blockBuilder, LikeMatcher.compile("helloX_world", Optional.of('X')));
-        LIKE_PATTERN.writeObject(blockBuilder, LikeMatcher.compile("foo%_bar"));
+        LIKE_PATTERN.writeObject(blockBuilder, DfaLikeMatcher.compile("helloX_world", Optional.of('X')));
+        LIKE_PATTERN.writeObject(blockBuilder, DfaLikeMatcher.compile("foo%_bar"));
+        Block block = blockBuilder.build();
+
+        LikeMatcher pattern = (LikeMatcher) LIKE_PATTERN.getObject(block, 0);
+        assertThat(pattern.getPattern()).isEqualTo("helloX_world");
+        assertThat(pattern.getEscape()).isEqualTo(Optional.of('X'));
+
+        pattern = (LikeMatcher) LIKE_PATTERN.getObject(block, 1);
+        assertThat(pattern.getPattern()).isEqualTo("foo%_bar");
+        assertThat(pattern.getEscape()).isEqualTo(Optional.empty());
+    }
+
+    @Test
+    public void testGetObjectForRegexLikeMatcher()
+    {
+        BlockBuilder blockBuilder = LIKE_PATTERN.createBlockBuilder(new PageBuilderStatus().createBlockBuilderStatus(), 10);
+        LIKE_PATTERN.writeObject(blockBuilder, RegexLikeMatcher.compile("helloX_world", Optional.of('X')));
+        LIKE_PATTERN.writeObject(blockBuilder, RegexLikeMatcher.compile("foo%_bar"));
         Block block = blockBuilder.build();
 
         LikeMatcher pattern = (LikeMatcher) LIKE_PATTERN.getObject(block, 0);
